Chg: refactoring the authentication module and removing redundant code blocks
